Troca float por decimal

+++ do_stuff.py
@@ -2,6 +2,8 @@ from gauss_seidel import gauss_seidel
 from lagrange import lagrange
 from lu_decomposition import lu_decomposition
 
+from decimal import Decimal
+
 def do():
     # Para ambos os sistemas, cria a matriz A, o vetor b e recebe os seus respectivos valores
     # Sistema 1
@@ -9,18 +11,18 @@ def do():
     b1 = []
     samples = int(input("Enter number of samples: "))
     for i in range(samples):
-        A1.append([float(x) for x in (input("Enter values of row %s in A1: " % str(i + 1)).split(" "))])
-    b1 = [float(x) for x in input("Enter values of b1: ").split(" ")]
+        A1.append([Decimal(x) for x in (input("Enter values of row %s in A1: " % str(i + 1)).split(" "))])
+    b1 = [Decimal(x) for x in input("Enter values of b1: ").split(" ")]
 
     # Sistema 2
     A2 = []
     b2 = []
     for i in range(samples):
-        A2.append([float(x) for x in (input("Enter values of row %s in A2: " % str(i + 1)).split(" "))])
-    b2 = [float(x) for x in input("Enter values of b2: ").split(" ")]
+        A2.append([Decimal(x) for x in (input("Enter values of row %s in A2: " % str(i + 1)).split(" "))])
+    b2 = [Decimal(x) for x in input("Enter values of b2: ").split(" ")]
 
     # Recebe o valor da precisão
-    eps = float(input("Enter precision: "))
+    eps = Decimal(input("Enter precision: "))
 
     x1 = lu_decomposition(A1, b1)
     x2 = gauss_seidel(A2, b2, eps)
